+++ R/diffusion.R
@@ -46,7 +46,6 @@
 #' @param maxiter number of iterations the optimser takes (default ==
 #'   \code{10000} for "nm" and \code{Inf} for "hj")
 #' @param opttol Tolerance for convergence (default == 1.e-06)
-#' @param na.rm Removes NA values
 #' 
 #' @return Returns an object of class \code{diffusion}, which contains:
 #' \itemize{
@@ -123,27 +122,22 @@ diffusion <- function(x, w = NULL, cleanlead = c(TRUE, FALSE), prew = NULL,
                       l = 2, cumulative = c(TRUE, FALSE), pvalreps = 0, 
                       eliminate = c(FALSE, TRUE), sig = 0.05, verbose = c(FALSE, TRUE),
                       type = c("bass", "gompertz", "gsgompertz", "weibull"),
-                      optim = c("nm", "hj"), maxiter = Inf, opttol = 1.e-06,
-                      na.rm = c(FALSE, TRUE)) {
+                      optim = c("nm", "hj"), maxiter = Inf, opttol = 1.e-06) {
 
   type <- match.arg(type, c("bass", "gompertz", "gsgompertz", "weibull"))
   optim <- match.arg(optim, c("nm", "hj"))
   
-  cleanlead <- cleanlead[1]
-  if (cleanlead == TRUE) {
-    x <- cleanzero(x)$x
-  }
-  
-  na.rm <- na.rm[1]
-  if (na.rm == TRUE) {
-    x <- removena(x)$x
-  }
-  
   cumulative <- cumulative[1]
   eliminate <- eliminate[1]
   verbose <- verbose[1]
+  cleanlead <- cleanlead[1] # note dependency in seqdiffusion plot
   
-  # Optimise parameters
+  if (cleanlead == TRUE) {
+    x <- cleanzero(x)$x
+  }
+  x <- cleanna(x)$x
+  
+    # Optimise parameters
   if (is.null(w)) {
 
     opt <- diffusionEstim(x, l, cumulative, prew, pvalreps, eliminate,
@@ -211,7 +205,7 @@ diffusionEstim <- function(x, l = 2, cumulative = c(FALSE, TRUE),
   
   if (eliminate == TRUE & pvalreps == 0){
     pvalreps <- 1000
-    warning("To eliminate parameters p-values must be estimated. Setting pvalreps = 1000.")
+    message("To eliminate parameters p-values must be estimated. Setting pvalreps = 1000.")
   }
   
   # Check botstrap repetitions (pvalreps)
@@ -267,7 +261,7 @@ diffusionEstim <- function(x, l = 2, cumulative = c(FALSE, TRUE),
         if (maxiter < 5000){
           #if (20*length(w.idx)^2 > 1500) {
           maxiter <- 5000
-          warning("Set maxiter to 5'000 for Naelder-Maed optimiser")
+          message("Set maxiter to 5'000 for better results with Naelder-Maed optimiser")
         } else if (maxiter == Inf) {
           maxiter <- 10000
         }
@@ -303,7 +297,7 @@ diffusionEstim <- function(x, l = 2, cumulative = c(FALSE, TRUE),
         
         if (maxiter < 100000){
           maxiter <- 100000
-          warning("Set maxiter to 100'000 for hj optimiser")
+          message("Set maxiter to 100'000 for hj optimiser")
         }
         
         # Hooker-Jeeves is slow but optimises tough stuff